+++ cpp/src/scheduler/event/StartUpEvent.h
@@ -17,6 +17,13 @@ public:
     explicit
     StartUpEvent(std::weak_ptr<Resource> resource)
         : Event(EventType::START_UP, std::move(resource)) {}
+
+    inline std::string
+    Dump() const override {
+        return "<StartUpEvent>";
+    }
+
+    friend std::ostream &operator<<(std::ostream &out, const StartUpEvent &event);
 };
 
 }