+++ cpp/src/scheduler/event/TaskTableUpdatedEvent.h
@@ -17,6 +17,13 @@ public:
     explicit
     TaskTableUpdatedEvent(std::weak_ptr<Resource> resource)
         : Event(EventType::TASK_TABLE_UPDATED, std::move(resource)) {}
+
+    inline std::string
+    Dump() const override {
+        return "<TaskTableUpdatedEvent>";
+    }
+
+    friend std::ostream &operator<<(std::ostream &out, const TaskTableUpdatedEvent &event);
 };